requirements.txt
streamlit>=1.38.0
openai>=1.40.0
python-dotenv>=1.0.1
pandas>=2.2.2
python-slugify>=8.0.4
requests>=2.32.3
